Added data.hide model detection
Changes to the data.hide model will now update chart appropriately.
Conversely, if the model is updated by clicking the legend items, the
model is updated and a digest is triggered via $apply.

The current implementation makes it impossible for a user to implement their
own onclick (or for that matter, their own legend). This will be addressed in
a future release.