Added expression attributes for click events
Instead of populating onclick functions in the model, events (specifically
data.onclick and legend.item.onclick) are now handled via attributes
(c3-data-click and c3-legend-click, respectively). See README for an example
and available expressions for each.

This is the "more Angular way" but also was necessary to properly handle events
within the correct scope, such as when the c3 model is changed to an identical
one according to angular.equals, but with new objects/arrays; this will not
trigger the model's $watch handler, and can result in the directive having a
stale model (which the parent scope's onclick handlers may update without the
directive responding). Also, the user's controller will need to wrap it's
callback in $applyAsync, which it shouldn't have to worry about.

A simple and likely-to-occur example showing why this change was necessary:

    // in my controller...

    $scope.chart = {
      data: {
        columns: [...],
        hide: []
      }
    };

    $scope.unhideData = function() {
      $scope.chart.hide = [];
    };

If $scope.chart.data.hide is already empty when this is called, then 'hide' will
change but because angular.equals([], []), the directive will not know about it
and its onclick callback will continue to operate on the old, stale version of
the hide object (meaning changes it makeswon't actually take effect). This is
avoided when using scope expressions because innocuous model changes won't
trigger $watch updates as before, but once a meaningful change occurs, it will
be handled within the directive, using the directive's scope.

Other changes include:

 - scope.c3 is used whenever possible, rather than any cached or closure'd
   model value; this a good practice in line with the fix above.

 - angular.merge is used instead of angular.extend to construct the config
   object, meaning that {legend: {...}} config parameters are now no longer
   overwritten (excluding, of course, legend.item.onclick).

 - A new config object is only created when needed (i.e. when the load API
   cannot be used), not on every $watch evaluation.

 - The internal toggleData function was renamed to toggleHide, which is more
   semantically correct.